[ARM] video: msm: Add support for MDP 3.1 (qsd8k)
Includes support for the PPP which does scale, rotate, blend, color-convert,
etc.

Also adds support for the LCDC output.

Signed-off-by: Dima Zavin <dima@android.com>

[ARM] video: mdp: add support for pre-multiplied alpha for mdp 3.1

Signed-off-by: Dima Zavin <dima@android.com>

HACK: [ARM] video: msm: lcdc: Set ebi1 clock rate to 128Mhz on resume

MDP clock rate is tied to the AXI, and setting the MDP clock doesn't work.
So, we request the ebi1 clock at the correct frequency whenever the screen
is on, which will force the MDP clock higher.

Signed-off-by: Dima Zavin <dima@android.com>

[ARM] msm: mdp: cleanup in probe corectly on failure

Change-Id: Ife63744371480048ae1523ec1185f9acd0e637bf
Signed-off-by: Dima Zavin <dima@android.com>

[ARM] video: msm: mdp: Set the ebi1 clock to 128mhz when performing blits

Thsi makes sure that the MDP (locked to AXI) and the AXI/ABI1 run at the
maximum frequency during blit operations.

Change-Id: Ifec2f4258c00146f478fc9941a261b89254bac0f
Signed-off-by: Dima Zavin <dima@android.com>

[ARM] video: msm: mdp_lcdc: Do not force EBI1 to max whenever the screen is on

Previously, when the frambuffer was located in the EBI, there would be visual
artifacts if the EBI1/AXI frequency was not forced to 128mhz. This should no
longer be required, and we should let the ARM9 scale EBI/AXI appropriately
during idle or static screen refreshes. The MDP blits will force it higher
when necessary.

Change-Id: Iace4a056abbe50dc7d1ddb70fe3636873e5cfaf3
Signed-off-by: Dima Zavin <dima@android.com>

[ARM] video: msm: Allow framebuffer to be configured in 32 bit mode

For now only accept XRGB_8888, RBG_888 and RGB_565

Signed-off-by: Rebecca Schultz Zavin <rebecca@android.com>

[ARM] video: msm: Allow users to request a larger x and yvirtual fb

As long as the total size fits inside the memory allocated to the framebuffer
users can request a larger size.  This change allows users to request
a triple buffered fb.

Signed-off-by: Rebecca Schultz Zavin <rebecca@android.com>

[ARM] video: msm: Correct use of resource in fb

The resource end is the last byte, add one to get the actual size.

Signed-off-by: Rebecca Schultz Zavin <rebecca@android.com>

[ARM] video: msm: mdp_lcdc: Don't update the dma config at init

Updating the dma config at init time leads to a race condition with the
framebuffer which hasn't configured the format yet.  Also if the dma config
of the bootloader and the kernel don't match this will cause artifacts.

Signed-off-by: Rebecca Schultz Zavin <rebecca@android.com>

[ARM] video: msm: mdp_lcdc: Correct lcdc timings

Signed-off-by: Rebecca Schultz Zavin <rebecca@android.com>

video: msm: Add support for 32 bit framebuffer packed BGR instead of RGB

This is the 32 bit format produced by the gpu.

Signed-off-by: Rebecca Schultz Zavin <rebecca@android.com>

ARM: msm: mdp: Catch blits that the mdp can't handle and return an error

Adds checks for the following two conditions:
	1.  ROI extends outside the area of the source or destination image
	2.  Source width or height is greater than 2048 (MDP bug)

Signed-off-by: Colin Cross <ccross@android.com>
Change-Id: I28d4dc46f5b157ebf7cacbbf5d00962fa2851c0e

ARM: msm: mdp: Disable pixel-repeat mode for scaling

Pixel-repeat mode was used for RGB->RGB blits, but it causes scaled text
to look bad.  Using the same FIR tables that we use for YUV blits looks
better.

Change-Id: I3041c9c42458da09b29489c5b2e8ac5b01f2c105
Signed-off-by: Colin Cross <ccross@android.com>

[ARM] msm: mdp: Disable SVI (sharpening) for RGB blits

Leaves SVI enabled for YUV frames (generally video).

Change-Id: If708e452896634c50ffa9026eb4f08c10819bc52
Signed-off-by: Colin Cross <ccross@android.com>

video: msm: Change the output to 565 instead of 888

Signed-off-by: Rebecca Schultz Zavin <rebecca@android.com>

[ARM] msm: mdp: Mask out old DMA destination bits when setting new ones

Change-Id: I3b9c15071bf30b93788a43ee931e5064cd337f78
Signed-off-by: Colin Cross <ccross@android.com>

[ARM] msm: mdp: Prevent framebuffer glitch during initialization

Holds a reference to the mdp_clk until lateinit, and moves the framebuffer
initialization to device_init.  The framebuffer lcdc driver will grab a
reference to mdp_clk, which prevents the clock from being disabled by
clock_late_init

Signed-off-by: Colin Cross <ccross@android.com>
Change-Id: I70d49b4c78b3cca8a52cf846b370d8ea7ffaa418

[ARM] msm: mdp: Prevent DMA lockups when switching bit depths

Forces the lcdc off and waits a frame vsync period to prevent the MDP DMA
engine crashing if the DMA bit depth is changed.  A sleep cannot be used
because lcdc_dma_start is called in irq context.

Change-Id: Ibdceede494a6edd8f49878419f5410cabafe0253
Signed-off-by: Colin Cross <ccross@android.com>

[ARM] msm: mdp: Print useful messages and BUG when the MDP locks up

Change-Id: Id930bc638ca86ae406c2714f5789d8f26363683a
Signed-off-by: Colin Cross <ccross@android.com>

[ARM] msm: mdp: Add more debugging output on mdp crashes

Signed-off-by: Colin Cross <ccross@android.com>
Change-Id: I4c188a702987a7a3baff6aab0eee77348e84a6ad

[ARM] msm: mdp: Add additional debugging info when blits fail

Change-Id: I5ddf43aef8b53d19c3a1cc160676ce2d0785914f