msm: trout/sapphire: move *_CPLD_BASE outside the vmalloc range
SAPPHIRE_CPLD_BASE and TROUT_CPLD_BASE were inside the range of memory
allocated by the kernel for virtual mappings, causing the IO mapping
to be overwritten if the kernel allocated the page.

To reproduce, allocate then deallocate all kernel memory.  The mapping
will be removed and the kernel will crash when any of the IO memory is
accessed (e.g. for changing phone brightness).

Change-Id: I49b3b8af5257ba77998e03d0ecff0f909ddc752c