msm: serial: Fix wakeup handling
Make arm/mach-msm use deferred irq disabling by using the default
handler.

Fix msm_serial and msm_serial_hs to not keep unhandled IRQs in the
wake state.

Increase the timeout on voltage loss to 2 seconds, because the
userspace was not getting enough time to set the LED without the
serial interrupt waking the phone.

Newer versions of Linux "disable" interrupts (including wakeup irqs)
during suspend, and reenable them on resume.  This works because it
uses deferred disabling, but the ARM code was still using disable as
a hardware mask_irq.

Due to the new handling, you can no longer leave unhandled AND disabled
IRQs as wake irqs, because they will always be pending, and keep the
system from sleeping.